requirements.txt
click
packaging
pyyaml
requests
semver
typing
